rename: AuthServiceProvicer -> AuthService
